+++ atdpy/src/lib/Codegen.ml
@@ -74,7 +74,14 @@ let to_camel_case s =
         Buffer.add_char buf c;
         start_word := false
   done;
-  Buffer.contents buf
+  let name = Buffer.contents buf in
+  if name = "" then "X"
+  else
+    (* Make sure we don't start with a digit. This happens with
+       generated identifiers like '_42'. *)
+    match name.[0] with
+    | 'A'..'Z' | 'a'..'z' | '_' -> name
+    | _ -> "X" ^ name
 
 (* Use CamelCase as recommended by PEP 8. *)
 let class_name env id =
@@ -498,7 +505,7 @@ let rec type_name_of_expr env (e : type_expr) : string =
       in
       sprintf "Tuple[%s]" (String.concat ", " type_names)
   | List (loc, e, an) ->
-      (match assoc_kind loc e an with
+     (match assoc_kind loc e an with
        | Array_list
        | Object_list _ ->
            sprintf "List[%s]"
@@ -515,7 +522,7 @@ let rec type_name_of_expr env (e : type_expr) : string =
   | Shared (loc, e, an) -> not_implemented loc "shared"
   | Wrap (loc, e, an) -> todo "wrap"
   | Name (loc, (loc2, name, []), an) -> py_type_name env name
-  | Name (loc, _, _) -> not_implemented loc "parametrized types"
+  | Name (loc, (_, name, _::_), _) -> assert false
   | Tvar (loc, _) -> not_implemented loc "type variables"
 
 let rec get_default_default
@@ -1212,10 +1219,16 @@ let run_file src_path =
     |> String.lowercase_ascii
   in
   let dst_path = dst_name in
-  let (atd_head, atd_module), _original_types =
+  let full_module, _original_types =
     Atd.Util.load_file
       ~annot_schema
-      ~expand:false ~inherit_fields:true ~inherit_variants:true src_path
+      ~expand:true (* monomorphization = eliminate parametrized type defs *)
+      ~keep_builtins:true
+      ~inherit_fields:true
+      ~inherit_variants:true
+      src_path
   in
+  let full_module = Atd.Ast.use_only_specific_variants full_module in
+  let (atd_head, atd_module) = full_module in
   let head = Python_annot.get_python_json_text (snd atd_head) in
   to_file ~atd_filename:src_name ~head atd_module dst_path